fixed some compile errors on osx

+++ src/games/Arena/BehaviorTree/PriorityNode.cpp
@@ -46,7 +46,7 @@ BEHAVIOR_STATUS PriorityNode::execute(void* agent)
 		else if (status == BT_FAILURE)
 		{
 			currentPosition++;
-			if (currentPosition == children.size())
+			if (currentPosition == (int)children.size())
 			{
 				currentPosition = -1;
 				return BT_FAILURE;
@@ -67,7 +67,7 @@ BEHAVIOR_STATUS PriorityNode::execute(void* agent)
 	{
 		currentPosition++;
 		currentlyRunningNode = children.at(currentPosition);
-		if (currentPosition == children.size()) //all of the children failed
+		if (currentPosition == (int)children.size()) //all of the children failed
 		{
 			currentPosition = -1;
 			return BT_FAILURE;
